Fix unit tests for retrieve(mapboxID:) details api
