+++ Sources/MapboxSearch/PublicAPI/Search Results/SearchCategorySuggestionImpl.swift
@@ -47,8 +47,6 @@ class SearchCategorySuggestionImpl: SearchCategorySuggestion, CoreResponseProvid
         self.batchResolveSupported = coreResult.action?.multiRetrievable ?? false
         self.categories = coreResult.categories
         self.categoryIDs = coreResult.categoryIDs
-        NSLog("@@ categories is \(coreResult.categories)")
-        NSLog("@@ category IDs is \(coreResult.categoryIDs)")
 
         self.descriptionText = coreResult.addressDescription
         self.estimatedTime = coreResult.estimatedTime
