+++ Tests/MapboxSearchTests/Legacy/SearchEngineTests.swift
@@ -485,194 +485,4 @@ class SearchEngineTests: XCTestCase {
         searchEngine.query = "random-query"
         XCTAssertEqual(searchEngine.query, "random-query")
     }
-
-    /// NOTE: Although this test uses separate fetches for each attribute set this is purely for testing coverage
-    /// purposes. It is recommended to request as many attribute sets as desired in one RequestOptions array.
-    /// Ex: You should use RetrieveOptions(attributeSets: [.visit, .photos]) in one select() call rather than two calls.
-    func testRetrieveDetailsFunction() throws {
-        let searchEngine = SearchEngine(apiType: .searchBox)
-        let delegate = SearchEngineDelegateStub()
-        searchEngine.delegate = delegate
-        let updateExpectation = delegate.updateExpectation
-
-        let searchOptions = SearchOptions(
-            limit: 100,
-            origin: CLLocationCoordinate2D(latitude: 38.902309, longitude: -77.029129),
-            filterTypes: [.poi]
-        )
-
-        searchEngine.search(query: "planet word", options: searchOptions)
-        wait(for: [updateExpectation], timeout: 200)
-        let suggestion = try XCTUnwrap(delegate.resolvedSuggestions?.first)
-
-        func fetchResult(for: SearchSuggestion, options: RetrieveOptions) throws -> SearchResult {
-            let successExpectation = delegate.successExpectation
-            searchEngine.select(suggestion: suggestion, options: options)
-            wait(for: [successExpectation], timeout: 200)
-            return try XCTUnwrap(delegate.resolvedResult)
-        }
-
-        let attributes = [
-            AttributeSet.basic,
-            .photos,
-            .venue,
-            .visit,
-        ]
-
-        let resultsByAttribute = try attributes.map { attributeSet in
-            let result = try fetchResult(for: suggestion, options: RetrieveOptions(attributeSets: [attributeSet]))
-
-            XCTAssertNotNil(result.metadata, "\(attributeSet) metadata should not be nil")
-
-            return (attributeSet, result)
-        }
-
-        XCTAssertNotNil(resultsByAttribute)
-        for (attribute, result) in resultsByAttribute {
-            let metadata = try XCTUnwrap(result.metadata)
-            switch attribute {
-            case .basic:
-                XCTAssertNotNil(metadata.primaryImage)
-                XCTAssertNil(metadata.otherImages)
-                XCTAssertNil(metadata.phone)
-                XCTAssertNil(metadata.website)
-                XCTAssertNil(metadata.reviewCount)
-                XCTAssertNil(metadata.averageRating)
-                XCTAssertNil(metadata.openHours)
-            case .photos:
-                XCTAssertNotNil(metadata.primaryImage)
-                XCTAssertNotNil(metadata.otherImages)
-                XCTAssertNil(metadata.phone)
-                XCTAssertNil(metadata.website)
-                XCTAssertNil(metadata.reviewCount)
-                XCTAssertNil(metadata.averageRating)
-                XCTAssertNil(metadata.openHours)
-            case .venue:
-                XCTAssertNotNil(metadata.primaryImage)
-                XCTAssertNil(metadata.otherImages)
-                XCTAssertNil(metadata.phone)
-                XCTAssertNil(metadata.website)
-                XCTAssertNotNil(metadata.reviewCount, "Review count failed for \(String(describing: result.mapboxId))")
-                XCTAssertNil(metadata.averageRating)
-                XCTAssertNil(metadata.openHours)
-            case .visit:
-                XCTAssertNotNil(metadata.primaryImage)
-                XCTAssertNil(metadata.otherImages)
-                XCTAssertNotNil(metadata.phone)
-                XCTAssertNotNil(metadata.website)
-                XCTAssertNil(metadata.reviewCount)
-                XCTAssertNil(metadata.averageRating)
-                XCTAssertNotNil(metadata.openHours)
-            }
-        }
-
-        XCTAssertNil(delegate.error)
-    }
-
-    func testRetrieveDetailsAllAttributes() throws {
-        let searchEngine = SearchEngine(apiType: .searchBox)
-        let delegate = SearchEngineDelegateStub()
-        searchEngine.delegate = delegate
-        let updateExpectation = delegate.updateExpectation
-
-        let searchOptions = SearchOptions(
-            limit: 100,
-            origin: CLLocationCoordinate2D(latitude: 38.902309, longitude: -77.029129),
-            filterTypes: [.poi]
-        )
-
-        searchEngine.search(query: "planet word", options: searchOptions)
-        wait(for: [updateExpectation], timeout: 200)
-        let suggestion = try XCTUnwrap(delegate.resolvedSuggestions?.first)
-
-        func fetchResult(for: SearchSuggestion, options: RetrieveOptions) throws -> SearchResult {
-            let successExpectation = delegate.successExpectation
-            searchEngine.select(suggestion: suggestion, options: options)
-            wait(for: [successExpectation], timeout: 200)
-            return try XCTUnwrap(delegate.resolvedResult)
-        }
-
-        let allAttributes = [
-            AttributeSet.basic,
-            .photos,
-            .venue,
-            .visit,
-        ]
-
-        let result = try fetchResult(for: suggestion, options: RetrieveOptions(attributeSets: allAttributes))
-
-        let metadata = try XCTUnwrap(result.metadata)
-        XCTAssertNil(metadata.averageRating)
-        XCTAssertNotNil(metadata.otherImages)
-        XCTAssertNotNil(metadata.openHours)
-        XCTAssertNotNil(metadata.phone)
-        XCTAssertNotNil(metadata.primaryImage)
-        XCTAssertNotNil(metadata.reviewCount, "Review count failed for \(String(describing: result.mapboxId))")
-        XCTAssertNotNil(metadata.website)
-        XCTAssertNil(delegate.error)
-    }
-
-    func testRetrieveDetailsByMapboxId() throws {
-        let searchEngine = SearchEngine(apiType: .searchBox)
-        let delegate = SearchEngineDelegateStub()
-        searchEngine.delegate = delegate
-        let updateExpectation = delegate.updateExpectation
-
-        let planetWordMapboxID = "dXJuOm1ieHBvaTo0ZTg2ZWFkNS1jOWMwLTQ3OWEtOTA5Mi1kMDVlNDQ3NDdlODk"
-        let detailsOptions = DetailsOptions(attributeSets: AttributeSet.allCases, language: "en")
-        searchEngine.retrieve(mapboxID: planetWordMapboxID, options: detailsOptions)
-        wait(for: [updateExpectation], timeout: 200)
-        let suggestion = try XCTUnwrap(delegate.resolvedSuggestions?.first)
-
-        func fetchResult(for: SearchSuggestion, options: RetrieveOptions) throws -> SearchResult {
-            let successExpectation = delegate.successExpectation
-            searchEngine.select(suggestion: suggestion, options: options)
-            wait(for: [successExpectation], timeout: 200)
-            return try XCTUnwrap(delegate.resolvedResult)
-        }
-
-        let allAttributes = [
-            AttributeSet.basic,
-            .photos,
-            .venue,
-            .visit,
-        ]
-
-        let result = try fetchResult(for: suggestion, options: RetrieveOptions(attributeSets: allAttributes))
-
-        let metadata = try XCTUnwrap(result.metadata)
-        XCTAssertNil(metadata.averageRating)
-        XCTAssertNotNil(metadata.otherImages)
-        XCTAssertNotNil(metadata.openHours)
-        XCTAssertNotNil(metadata.phone)
-        XCTAssertNotNil(metadata.primaryImage)
-        XCTAssertNil(metadata.reviewCount, "Review count failed for \(String(describing: result.mapboxId))")
-        XCTAssertNotNil(metadata.website)
-        XCTAssertNil(delegate.error)
-    }
-
-    func testRetrieveDetailsByMapboxIDForPOIWithEmptyOptions() throws {
-        let searchEngine = SearchEngine(apiType: .searchBox)
-        let delegate = SearchEngineDelegateStub()
-        searchEngine.delegate = delegate
-        let successExpectation = delegate.successExpectation
-
-        let planetWordMapboxID = "dXJuOm1ieHBvaTo0ZTg2ZWFkNS1jOWMwLTQ3OWEtOTA5Mi1kMDVlNDQ3NDdlODk"
-        searchEngine.retrieve(mapboxID: planetWordMapboxID, options: nil)
-        wait(for: [successExpectation], timeout: 200)
-        XCTAssertNil(delegate.error)
-
-        let result = try XCTUnwrap(delegate.resolvedResult)
-        let metadata = try XCTUnwrap(result.metadata)
-
-        XCTAssertNil(metadata.averageRating)
-        XCTAssertNotNil(metadata.primaryImage)
-        XCTAssertNil(metadata.reviewCount)
-
-        // In this test no attribute sets were provided and these fields are nil
-        XCTAssertNil(metadata.otherImages)
-        XCTAssertNil(metadata.openHours)
-        XCTAssertNil(metadata.phone)
-        XCTAssertNil(metadata.website)
-    }
 }
